requirements.txt
aiofiles==0.7.0
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
fastapi==0.68.0
h11==0.12.0
idna==3.2
Jinja2==3.0.1
joblib==1.0.1
MarkupSafe==2.0.1
nltk==3.6.2
numpy==1.21.1
oauthlib==3.1.1
pandas==1.3.1
pickle-mixin==1.0.2
pydantic==1.8.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-multipart==0.0.5
pytz==2021.1
regex==2021.7.6
requests==2.26.0
requests-oauthlib==1.3.0
scikit-learn==0.24.2
scipy==1.7.1
six==1.16.0
sklearn==0.0
starlette==0.14.2
stopwords==1.0.0
threadpoolctl==2.2.0
tqdm==4.62.0
tweepy==3.10.0
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
